Moved browse and feature use cases into coffee package
