+++ domain/src/main/java/com/alexzh/imbarista/domain/interactor/coffee/favourite/RemoveCoffeeFromFavourite.kt
@@ -1,4 +1,4 @@
-package com.alexzh.imbarista.domain.interactor.favourite
+package com.alexzh.imbarista.domain.interactor.coffee.favourite
 
 import com.alexzh.imbarista.domain.executor.PostExecutionThread
 import com.alexzh.imbarista.domain.interactor.CompletableUseCase
@@ -6,19 +6,19 @@ import com.alexzh.imbarista.domain.repository.CoffeesRepository
 import io.reactivex.Completable
 import java.lang.IllegalArgumentException
 
-class AddCoffeeToFavourites constructor(
+class RemoveCoffeeFromFavourite constructor(
     private val coffeesRepository: CoffeesRepository,
     postExecutionThread: PostExecutionThread
-) : CompletableUseCase<AddCoffeeToFavourites.Param>(postExecutionThread) {
+) : CompletableUseCase<RemoveCoffeeFromFavourite.Param>(postExecutionThread) {
 
     override fun buildCompletableUseCase(params: Param?): Completable {
         if (params == null) throw IllegalArgumentException("Param can't be null")
-        return coffeesRepository.addCoffeeToFavourites(params.coffeeId)
+        return coffeesRepository.removeCoffeeFromFavourites(params.coffeeId)
     }
 
     data class Param(val coffeeId: Long) {
         companion object {
-            fun forCoffee(id: Long): Param {
+            fun forCoffee(id: Long) : Param {
                 return Param(id)
             }
         }
